Flytt versjonert søknad til journalpost pakke, bruk felles exceptionhandling
